Refactor. Changed time-related settings to longs for consistency.
Since were not measuring time at a grainularity smaller that millis, it is misleading to have settings that specify times using floats.

+++ FittsInstance.pde
@@ -29,7 +29,7 @@ class FittsInstance {
 
   public void update(long frameTimeMillis, HashMap<Action, Float> readings) {
     float speed = readings.get(Action.RIGHT) - readings.get(Action.LEFT);
-    float newX = state.relativeCursorX + speed*(frameTimeMillis/settings.travelTimeMillis);
+    float newX = state.relativeCursorX + speed*(((float)frameTimeMillis)/settings.travelTimeMillis);
     newX = max(-1.0, newX);
     newX = min(newX, 1.0);
